Move IndentedWriter from outer to top-level class
