+++ Cql/CodeGeneration.NET/IAddIndentMutable.cs
@@ -1,0 +1,6 @@
+﻿namespace Hl7.Cql.CodeGeneration.NET;
+
+internal interface IAddIndentMutable<out TSelf>
+{
+    TSelf AddIndent(int addIndent = 1);
+}
